+++ hahomematic/platforms/entity.py
@@ -422,6 +422,15 @@ class BaseParameterEntity(Generic[ParameterT, InputParameterT], BaseEntity):
                 channel_address=channel_address, parameter=parameter
             ),
         )
+        self._is_un_ignored: Final[
+            bool
+        ] = self._central.parameter_visibility.parameter_is_un_ignored(
+            device_type=self._device.device_type,
+            channel_no=self._channel_no,
+            paramset_key=self._paramset_key,
+            parameter=self._parameter,
+            custom_only=True,
+        )
         self._value: ParameterT | None = None
         self._last_updated: datetime = INIT_DATETIME
         self._last_refreshed: datetime = INIT_DATETIME
@@ -465,6 +474,11 @@ class BaseParameterEntity(Generic[ParameterT, InputParameterT], BaseEntity):
     def is_unit_fixed(self) -> bool:
         """Return if the unit is fixed."""
         return self._raw_unit != self._unit
+
+    @config_property
+    def is_un_ignored(self) -> bool:
+        """Return if the parameter is un ignored."""
+        return self._is_un_ignored
 
     @config_property
     def max(self) -> ParameterT: