+++ changelog.md
@@ -1,3 +1,8 @@
+# Version 2024.1.6 (2024-01-11)
+
+- Fix unignore doc and improve unignore tests
+- Move unignore check to entity
+
 # Version 2024.1.5 (2024-01-09)
 
 - Remove effects from HmIP-RGBW when in PWM mode